chore(web): move logger to utils/
